% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_utils.R
\name{admin_match}
\alias{admin_match}
\title{Match admin region}
\usage{
admin_match(admin = NULL, country = NULL, quiet = FALSE)
}
\arguments{
\item{admin}{Character for admin region. Some fuzzy logic will be used to 
match. If not provided then no seasonality is introduced. Default = NULL}

\item{country}{Character for country within which admin2 is in. 
Default = NULL}

\item{quiet}{Whether function should be quiet. If FALSE (default) the 
returned country and admin are printed.}
}
\value{
Row number of the matching region in 
  \code{magenta::admin_units_seasonal}
}
\description{
Match admin region
}
\keyword{internal}
