% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-state-create.R
\name{barcode_parms_create}
\alias{barcode_parms_create}
\title{Barcode parameter list creation}
\usage{
barcode_parms_create(num_loci = 24, ibd_length = 1, plaf = rep(0.5,
  24), prob_crossover = rep(0.5, 24), starting_ibd = 0)
}
\arguments{
\item{num_loci}{Number of loci. Default = 24}

\item{ibd_length}{If we are simulating IBD dynamics, each loci is now 
represented by a bitset of ibd_length. Thus ibd_length needs to be long
enough to ensure that as new identity relationships occur, i.e. 
an importation barcode will be a new identity. e.g. If your population is
1000, we may expect at 80% prevalence, with a mean COI of 3 we will need 
2400 different identities, i.e. 2^ibd_length > 2400. However, keep in mind
importations as these need to be continually new, i.e. if we are simulating
for 30 years, with 3 importations a day, then we will need at least length
to ensure that 2^ibd_length > 2400 + (30*365*3). This will probably be
automatically calculated in the future. If we are not ding IBD, then this
should be 1, which is the defalt.}

\item{plaf}{Vector of population level allele frequencies for the barcode. 
Default = rep(0.5, barcode_length)}

\item{prob_crossover}{Vector of probabilities for crossover events for the 
barcode. Default = rep(0.5, barcode_length)}

\item{starting_ibd}{Starting ibd. Default = 0}
}
\description{
\code{barcode_parms_create} creates list detialing the barcode/genetic
  parameters for the model
}
