% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bitsToInt}
\alias{bitsToInt}
\title{Function to convert vector of bools representing up to a 32 bit number into the corresponding integer}
\usage{
bitsToInt(x)
}
\arguments{
\item{x}{Vector of logicals}
}
\description{
Function to convert vector of bools representing up to a 32 bit number into the corresponding integer
}
