% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{COI_df_create}
\alias{COI_df_create}
\title{Create COI dataframe}
\usage{
COI_df_create(sim_save, groupvars = c("Age_Bin", "State"),
  barcodes = FALSE, mpl = Model_Param_List_Create(), ibd = 0,
  nl = 24, n = Inf, sample_group = c("T", "D", "A", "U"), reps = 1)
}
\arguments{
\item{sim_save}{Output of a simulation save}

\item{groupvars}{Grouping vars for summarySE. Default = c("Age_Bin","State")}

\item{barcodes}{Boolean whether to return tabled barcodes. Default = FALSE}
}
\description{
Create COI dataframe
}
