% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_param_lists.R
\name{drug_create_asaq}
\alias{drug_create_asaq}
\title{ASAQ Drug Create}
\usage{
drug_create_asaq()
}
\description{
ASAQ Drug Create
}
\note{
We have curves of the longest and shortest duration of ASAQ prophylaxis from 
Bretscher et al. However, we have multiple types of partner drug resistance. 
At the moment we will say that any loci associated with AQ resistance
yield the resistant curve but it may be better to have a distinct curve for each
AQ resistant genotype with the T0.5 aligned to the prob of lpf.
}
