% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_param_lists.R
\name{drug_create_asmq}
\alias{drug_create_asmq}
\title{ASMQ Drug Create}
\usage{
drug_create_asmq()
}
\description{
Create list describing parameters for ASMQ drug efficacy and prophlyaxis
}
\note{
Mefloquine params for prophylaxis not well known.
}
