% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_param_lists.R
\name{drug_create_default_no_resistance}
\alias{drug_create_default_no_resistance}
\title{Perfect Drug Create}
\usage{
drug_create_default_no_resistance()
}
\description{
Perfect Drug Create
}
\note{
Perfect Efficacy Drug. Used as default to match deterministic model easily
}
