% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_param_lists.R
\name{drug_create_dhappq}
\alias{drug_create_dhappq}
\title{DHA-PPQ Drug Create
Create list describing parameters for DHA-PPQ drug efficacy and prophlyaxis}
\usage{
drug_create_dhappq()
}
\description{
DHA-PPQ Drug Create
Create list describing parameters for DHA-PPQ drug efficacy and prophlyaxis
}
