% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_param_lists.R
\name{drug_list_create}
\alias{drug_list_create}
\title{Create drug list}
\usage{
drug_list_create(
  resistance_flag = FALSE,
  number_of_resistance_loci = 2,
  artemisinin_loci = c(0),
  cost_of_resistance = c(0.99, 0.99),
  absolute_fitness_cost_flag = FALSE,
  epistatic_logic = NULL,
  number_of_drugs = 1,
  drugs = list(drug_create_default_no_resistance()),
  mft_flag = FALSE,
  res_diag_flag = FALSE,
  time_res_diag = 999999,
  res_diag_cov = 1,
  temporal_cycling = -1,
  sequential_cycling = -1,
  sequential_update = 3,
  drug_choice = 0,
  partner_drug_ratios = matrix(rep(1/number_of_drugs, number_of_drugs), nrow = 1)
)
}
\arguments{
\item{resistance_flag}{Boolean are we simulating resistance}

\item{number_of_resistance_loci}{Numeric for number of res. loci}

\item{artemisinin_loci}{Numerics for barcode positions that confer artemisinin 
resistance}

\item{cost_of_resistance}{Numeric vector for fitness cost of each resistance loci.}

\item{absolute_fitness_cost_flag}{Boolean are we simulating fitness costs as
absolute (i.e. impactive onward transmission chance) or relative (resistant
strains have decreases chance of being onwardly transmitted in mixed 
infections). Default = FALSE.}

\item{epistatic_logic}{Is there compensatory relationships. i.e. what loci 
need to be true for resistance cost to exist. Default of NULL means that 
this becomes seq_len(number_of_resistance_loci), i.e. only dependent on 
their own loci. (TODO: Change this to be a list of length norl)}

\item{number_of_drugs}{Numeric for number of drugs used}

\item{drugs}{List of drugs that are being used, with each list element being
created by \code{drug_create}.}

\item{mft_flag}{Boolean are we doing mft}

\item{res_diag_flag}{Boolean are we doing resistance diagnostics}

\item{time_res_diag}{Integer time of introducing resistance diagnostics, years.}

\item{res_diag_cov}{Double resistance diagnostics coverage, proportion}

\item{temporal_cycling}{Numeric for when in years a drug switch occurs}

\item{sequential_cycling}{Numeric for what perc. treatment failure before switch}

\item{sequential_update}{How long does it take in years for sequential to be implemented}

\item{drug_choice}{What's the default drug choice to begin. Default = 0}

\item{partner_drug_ratios}{Numeric vector OR array for ratio of first line drugs used - changed to array over multiple years.}
}
\description{
List for simulating drug usage for resistance/mft variables
}
