% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_table}
\alias{drug_table}
\title{Drug Efficacy by Genotype Table}
\format{
A data.frame of drug efficacies by genotype
}
\usage{
drug_table
}
\description{
Table of probaility of late parasitological failure (28-day treatment failure)
 per genoytpe
}
\keyword{datasets}
