% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_state_create.R
\name{generate_default_model}
\alias{generate_default_model}
\title{Odin generator function}
\usage{
generate_default_model(ft, age, dat, generator, dde = TRUE)
}
\arguments{
\item{ft}{Treatment seeking numeric}

\item{age}{Vector of ages}

\item{dat}{Output of \code{equilibrium_init_create}}

\item{generator}{generator for model.}

\item{dde}{Use dde in soliving eventual odin. Default = FALSE}
}
\description{
\code{generate_default_model} creates the user required generator for
the odin model.
}
\keyword{internal}
