% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmap_ordered_binary_plot}
\alias{heatmap_ordered_binary_plot}
\title{Plot ordered heatmap}
\usage{
heatmap_ordered_binary_plot(
  sim_save,
  years,
  EIR,
  ordered = TRUE,
  save_path = NULL
)
}
\arguments{
\item{sim_save}{Saved output of simulation}

\item{years}{Lenth of simulation}

\item{EIR}{Numeric for simulation EIR}

\item{ordered}{Boolean for oredring heatmap. Default = TRUE}

\item{save_path}{File path where tiff is saved to. Default = NULL

\code{heatmap_ordered_binary_plot}}
}
\description{
\code{heatmap_ordered_binary_plot} creates a heatmap of all human barcodes, 
with the most common barcode grouped at the bottom
}
\keyword{internal}
