% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{importations}
\alias{importations}
\title{Importation data}
\format{A list of length 18

  \code{importations}: A list of length 18, with each list representing a year. In each
  year is then a further 2 lists which are:
    \itemize{
      \item incidence: Proportions of incidence that originated from other admin units, i.e.
      individuals in admin i that moved to admin j and then returned with an infection acquired
      while in admin j.  
      \item mosquitoFOI: Proportion of the force of infection towards mosquitoes that originated
      from outside admin units, i.e. the proportion of mosquitoes that are infected from infected
      individuals who travelled from admin j into admin i. 
    }}
\usage{
importations
}
\description{
Importation data for admin units in admin_units_seasonal
}
\keyword{datasets}
