% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{intervention_grab}
\alias{intervention_grab}
\title{Intervention grab}
\usage{
intervention_grab(country, admin, year_range, final_itn_cov = NULL,
  final_irs_cov = NULL, final_ft = NULL)
}
\arguments{
\item{final_ft}{}
}
\description{
Grabs ITN, IRS  ft from database
}
