% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irs_2000_2015}
\alias{irs_2000_2015}
\alias{irs_2010_2015}
\title{IRS for 2000 to 2015}
\format{A dataframe of 5 elements:

  \code{$irs_2000_2015}: A dataframe of admin units and their seasonal parameters
    \itemize{
      \item intervention: String stating IRS
      \item country: Country string
      \item admin: Admin string
      \item year: Numeric year
      \item value: Value for IRS coverage
    }}
\usage{
irs_2000_2015
}
\description{
IRS data for SSSA for 2000 to 2015
}
\keyword{datasets}
