% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{itn_2000_2015}
\alias{itn_2000_2015}
\title{ITN for 2000 to 2015}
\format{
A dataframe of 5 elements:

  \code{itn_2000_2015}: A dataframe of admin units and their seasonal parameters
    \itemize{
      \item intervention: String stating ITN
      \item country: Country string
      \item admin: Admin string
      \item year: Numeric year
      \item value: Value for ITN coverage
    }
}
\usage{
itn_2000_2015
}
\description{
ITN data for SSSA for 2000 to 2015
}
\keyword{datasets}
