% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixing_matrix}
\alias{mixing_matrix}
\title{Mixing Matrix}
\format{A matrix of 576x576:

  \code{mixing_matrix}: A matrix of numbers between 0 and 1 showing the proportion of casese
  due to importations by each other admin region. Along a row for a region, the number is the 
  proportion of infections due to importations from each region.}
\usage{
mixing_matrix
}
\description{
Importation data for admin units in admin_units_seasonal
}
\keyword{datasets}
