% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{mu_fv_create}
\alias{mu_fv_create}
\title{Create mosquito death and feeding vector}
\usage{
mu_fv_create(
  eqInit,
  ft,
  itn_cov,
  irs_cov,
  years,
  int_times = NULL,
  full = FALSE,
  odin_model = system.file("extdata/odin_itn_irs.R", package = "magenta")
)
}
\arguments{
\item{eqInit}{Equilibrium Inititial Conditions}

\item{ft}{Annual Treatment Coverage vector}

\item{itn_cov}{Annual ITN Coverage vector}

\item{irs_cov}{Annual IRS Coverage vector}

\item{years}{Numeric for total years}

\item{int_times}{Timing of intervention changes. Default = NULL, which will 
assume they are 365 days apart}

\item{full}{Boolean to return the whole deterministic output or just mosquito
death rate and anthrophagy. Default = FALSE}

\item{odin_model}{File path to odin model file. Default is the itns and irs
model included with magenta.}
}
\description{
Grabs spatial incidence and mosquitoFOI vector from database
}
\keyword{internal}
