% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{nmf_list_create}
\alias{nmf_list_create}
\title{Create nmf list}
\usage{
nmf_list_create(nmf_flag = FALSE, mean_nmf_frequency = c(247.63, 232.63,
  235.94, 259.79, 298.94, 360.32, 389.13, 446.76, 521.43, 525.94, 475.26,
  425.41, 397.17, 361.03), nmf_age_brackets = c(-0.1, 365, 730, 1095, 1460,
  1825, 2555, 3285, 4015, 4745, 5475, 7300, 9125, 10950, 36850),
  prob_of_testing_nmf = 0.5)
}
\arguments{
\item{prob_of_testing_nmf}{}
}
\description{
List for simulating non malarial fever
}
