% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{Param_List_Simulation_Create}
\alias{Param_List_Simulation_Create}
\title{Parameter List creation for MAGENTA simulation}
\usage{
Param_List_Simulation_Create(N = 10000, years = 1, eqSS)
}
\arguments{
\item{N}{Population size. Default = 1e4}

\item{years}{Length of simulation. Default = 1}

\item{eqSS}{Output of \code{Equilibrium_Steady_State_Create}}
}
\description{
\code{Param_List_Simulation_Create} creates suitable parameter list for
\code{Simulation_R}
}

