% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{param_list_simulation_finalizer_create}
\alias{param_list_simulation_finalizer_create}
\title{Parameter List creation for magenta simulation finalizer}
\usage{
param_list_simulation_finalizer_create(statePtr)
}
\arguments{
\item{statePtr}{Pointer for current model state as return by \code{simulation_R}$Ptr}
}
\description{
\code{param_list_simulation_finalizer_create} creates suitable parameter list for
\code{simulation_R} for free memory used by a simulation
}
