% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{param_list_simulation_get_create}
\alias{param_list_simulation_get_create}
\title{Parameter List creation for magenta simulation getting (saving to disk)}
\usage{
param_list_simulation_get_create(statePtr)
}
\arguments{
\item{statePtr}{Pointer for current model state as return by \code{simulation_R}$Ptr}
}
\description{
\code{param_list_simulation_get_create} creates suitable parameter list for
\code{simulation_R} for continuing a simulation from memory within the active
session.
}
