% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{param_list_simulation_saved_init_create}
\alias{param_list_simulation_saved_init_create}
\title{Parameter List creation for loading saved magenta simulation}
\usage{
param_list_simulation_saved_init_create(savedState)
}
\arguments{
\item{savedState}{Saved state generated by 
\code{simulation_R} when provided with a 
\code{param_list_simulation_get_create} parameter list}
}
\description{
\code{param_list_simulation_saved_init_create} creates 
suitable parameter list for \code{simulation_R} for continuing a 
simulation from memory within the active session.
}
