% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{Param_List_Simulation_Update_Create}
\alias{Param_List_Simulation_Update_Create}
\title{Parameter List creation for MAGENTA simulation updating}
\usage{
Param_List_Simulation_Update_Create(years = 1, statePtr)
}
\arguments{
\item{years}{Length of simulation. Default = 1}

\item{statePtr}{Pointer for current model state as return by \code{Simulation_R}$Ptr}
}
\description{
\code{Param_List_Simulation_Update_Create} creates suitable parameter list for
\code{Simulation_R} for continuing a simulation from memory within the active
session.
}

