% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{plaf_matrix_check}
\alias{plaf_matrix_check}
\title{plaf matrix check}
\usage{
plaf_matrix_check(matrix, years)
}
\arguments{
\item{years}{Years desired}

\item{plaf}{plaf matrix}
}
\description{
checks length formatting etc
}
