% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe_utils.R
\name{pop_strains_df}
\alias{pop_strains_df}
\title{Creates data frame containing the strain genetics of the population using 
the universe ptr. This data.frame is then passed to \code{COI_df_create}.}
\usage{
pop_strains_df(statePtr, sample_size = 0, sample_states = 0:5,
  ibd = FALSE, seed, nl = 24, big_mat_test = TRUE)
}
\arguments{
\item{statePtr}{statePtr}

\item{sample_size}{sample_size}

\item{sample_states}{sample_states}

\item{ibd}{ibd boolean.}
}
\description{
Creates data frame containing the strain genetics of the population using 
the universe ptr. This data.frame is then passed to \code{COI_df_create}.
}
