% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{population_get_genetics_ibd_df_n}
\alias{population_get_genetics_ibd_df_n}
\title{Returns the population's parasite genetics for ibd style summarised by pibd for given sample size and state}
\usage{
population_get_genetics_ibd_df_n(paramList)
}
\arguments{
\item{paramList}{paramList containing statePtr, sample_size, and sample_states}
}
\value{
list of population information
}
\description{
Returns the population's parasite genetics for ibd style summarised by pibd for given sample size and state
}
