% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Sample_COI}
\alias{Sample_COI}
\title{Sample x times from saved simulation and produce age COI data}
\usage{
Sample_COI(sim.save, sample_size, age_densities, age_breaks = seq(0, 90 * 365,
  2 * 365), reps)
}
\arguments{
\item{sim.save}{Saved output of simulation}

\item{sample_size}{Numeric for sample size}

\item{age_densities}{Densitvy vector for age distribution required}

\item{age_breaks}{Corresponding vector of age breaks for the density}

\item{reps}{How many samples are made}
}
\description{
\code{Sample_COI} samples from a human save according to some age distirbution
}
