% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Simulation_Finalizer_cpp}
\alias{Simulation_Finalizer_cpp}
\title{Returns whole model to R in series of nested lists}
\usage{
Simulation_Finalizer_cpp(param_list)
}
\arguments{
\item{param_list}{parameter list generated with \code{param_list_simulation_finalizer_create}}
}
\value{
list of 1 confirming finalizer has finished
}
\description{
Returns whole model to R in series of nested lists
}
