% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Simulation_Get_cpp}
\alias{Simulation_Get_cpp}
\title{Returns whole model to R in series of nested lists}
\usage{
Simulation_Get_cpp(paramList)
}
\arguments{
\item{paramList}{parameter list generated with \code{Param_List_Simulation_Get_Create}}
}
\value{
list of 4 lists with the entire model state
}
\description{
Returns whole model to R in series of nested lists
}
