% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Simulation_Init_cpp}
\alias{Simulation_Init_cpp}
\title{Creates initial model simulation using paramter list provided}
\usage{
Simulation_Init_cpp(paramList)
}
\arguments{
\item{paramList}{parameter list generated with \code{Param_List_Simulation_Init_Create}}
}
\value{
list with ptr to model state and loggers describing the current model state
}
\description{
Creates initial model simulation using paramter list provided
}
