% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{simulation_R}
\alias{simulation_R}
\title{simulation_R function}
\usage{
simulation_R(param_list, seed)
}
\arguments{
\item{param_list}{paramlist passed from 
\code{param_list_simulation_init_create} or from 
\code{param_list_simulation_update_create}}

\item{seed}{Seed for the simulation}
}
\description{
This function triggers the main magenta simulation from the R side
}
