% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Simulation_Saved_Init_cpp}
\alias{Simulation_Saved_Init_cpp}
\title{Creates initial model simulation using a saved model state}
\usage{
Simulation_Saved_Init_cpp(param_list)
}
\arguments{
\item{param_list}{parameter list generated with \code{Param_List_Simulation_Get_Create}}
}
\value{
list with ptr to model state and loggers describing the current model state
}
\description{
Creates initial model simulation using a saved model state
}
