% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_state_create.R
\name{spl_create}
\alias{spl_create}
\title{Spatial Parameter List Creation}
\usage{
spl_create(
  spatial_type = 0,
  human_importation_rate_vector = NULL,
  mosquito_imporation_rate_vector = NULL,
  cotransmission_freq_vector = rep(1, 10000),
  oocyst_freq_vector = rep(1, 10000),
  num_human_infs = 0,
  num_mos_infs = 0,
  plaf = rep(0.5, 24),
  island_imports_plaf_linked_flag = FALSE
)
}
\arguments{
\item{spatial_type}{Numeric for spatial type. Default = 0}

\item{human_importation_rate_vector}{Vector of human importations into this
region. Default = NULL}

\item{mosquito_imporation_rate_vector}{Vector of mosquito infection 
importations into this region. Default = NULL}

\item{cotransmission_freq_vector}{Vector of size of cotransmissions. This is 
preallocated here and regenerated every simulation update length.
Default = rep(1, 10000)}

\item{oocyst_freq_vector}{Vector of number of ooocysts. This is 
preallocated here and regenerated every simulation update length.
Default = rep(1, 10000)}

\item{num_human_infs}{Number of human infections from last day. Required for
metapopulation. Default = 0}

\item{num_mos_infs}{Number of mosquito infections from last day. Required for
metapopulation. Default = 0}

\item{plaf}{Population Level Frequency. Default = rep(0.5, 24)}

\item{island_imports_plaf_linked_flag}{Boolean. Whether imported barcodes for
island model are drawn dependent on other sites. Currently, if TRUE, imported
barcodes will either be all 0 or will be 1 at all loci where plaf > 0, 
if \code{rbinom(1, 1, plaf[plaf>0][1])} is TRUE. I.e. the first loci greater 
than 0 is used to determine if the incoming barcode is 1 at all sites where 
plaf is greater than 0.}
}
\description{
\code{spl_create} creates the user required spatial
list to be passed to the initialisation function if needed.
}
