% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{spl_matrix_check}
\alias{spl_matrix_check}
\title{Spatial matrix grab}
\usage{
spl_matrix_check(matrix, years)
}
\arguments{
\item{matrix}{Spatial matrix for 2000 to 2017}

\item{years}{Years desired}
}
\description{
Grabs spatial incidence and mosquitoFOI matrix from database
}
