% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarySE}
\alias{summarySE}
\title{Create summary statistics df}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{Dataframe to be summarised}

\item{measurevar}{Character for measure variable}

\item{groupvars}{Characters of grouping variables}

\item{na.rm}{Boolean to remove NAs.}

\item{conf.interval}{numeric for CI}

\item{.drop}{drop within ddply}
}
\description{
Create summary statistics df
}
