% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarySE_mean_only}
\alias{summarySE_mean_only}
\title{Create summary statistics df}
\usage{
summarySE_mean_only(data = NULL, measurevar = "COI",
  groupvars = c("Age_Bin", "State"))
}
\arguments{
\item{data}{Dataframe to be summarised}

\item{measurevar}{Character for measure variable}
}
\description{
Create summary statistics df
}
