% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bitsToInt}
\alias{bitsToInt}
\title{Function to convert vector of bools representing up to a 32 bit 
number into the corresponding integer}
\usage{
bitsToInt(x, big_endian = FALSE)
}
\arguments{
\item{x}{Vector of raw/logicals}

\item{big_endian}{Is big_endian. Default = FALSE}
}
\description{
Function to convert vector of bools representing up to a 32 bit 
number into the corresponding integer
}
\keyword{internal}
