% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{COI_age_plot_sample_x}
\alias{COI_age_plot_sample_x}
\title{Plot COI against age from random sample}
\usage{
COI_age_plot_sample_x(Sample_COI_out, x, span = 0.6, ylimmax = NULL,
  xlimmax = NULL, max_coi = 25)
}
\arguments{
\item{Sample_COI_out}{Output of Sample_COI}

\item{x}{Which sample to plot}

\item{span}{Smoothing parameter for loess}

\item{ylimmax}{ylim max}

\item{xlimmax}{xlim max}

\item{max_coi}{max_coi}
}
\description{
\code{Sample_COI} samples from a human save according to some age distirbution
}
