% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe_utils.R
\name{COI_df_create}
\alias{COI_df_create}
\title{Create COI dataframe}
\usage{
COI_df_create(df, groupvars = c("age_bin", "clinical"),
  breaks = c(-0.001, 5, 15, 100.1), barcodes = FALSE, ibd = 0,
  nl = 24, n = Inf, reps = 1, mean_only = TRUE)
}
\arguments{
\item{groupvars}{Grouping vars for summarySE. Default = c("Age_Bin","State")}

\item{barcodes}{Boolean whether to return tabled barcodes. Default = FALSE}

\item{sim_save}{Output of a simulation save}
}
\description{
Create COI dataframe
}
