% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-ordered-binary-plot.R
\name{Convert_Barcode_Vectors}
\alias{Convert_Barcode_Vectors}
\title{Convert human barcodes to numerics}
\usage{
Convert_Barcode_Vectors(sim.save)
}
\arguments{
\item{sim.save}{Saved output of simulation

\code{Convert_Barcode_Vectors}}
}
\description{
\code{Convert_Barcode_Vectors} converts human barcode vectors to nums
}
