% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{convert_barcode_vectors}
\alias{convert_barcode_vectors}
\title{Convert human barcodes to numerics}
\usage{
convert_barcode_vectors(
  sim_save,
  ID,
  sub_patents_included = TRUE,
  ibd = FALSE,
  nl = 24,
  COI_type = "pcr_imperial"
)
}
\arguments{
\item{sim_save}{Saved output of simulation}

\item{ID}{Vector of detection immunities for the population}

\item{sub_patents_included}{Boolean as to whether subpatents are included. 
Default = TRUE}

\item{ibd}{Boolean for IBD simulations. Default = FALSE}

\item{nl}{Numeric for number of loci. Default = 24}

\item{COI_type}{String for type of COI calculation. Default = "pcr_imperial"

\code{convert_barcode_vectors}}
}
\description{
\code{convert_barcode_vectors} converts human barcode vectors to nums and 
  calculated COIs.
}
