% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_utils.R
\name{country_seasonal_profiles}
\alias{country_seasonal_profiles}
\title{Plot country seasonal profiles}
\usage{
country_seasonal_profiles(country)
}
\arguments{
\item{country}{Character for country}
}
\description{
Plot country seasonal profiles
}
\keyword{internal}
