% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{drug_list_create}
\alias{drug_list_create}
\title{Create drug list}
\usage{
drug_list_create(resistance_flag = FALSE, number_of_drugs = 2,
  number_of_resistance_loci = 3, resistance_costs = c(0.99, 0.99,
  0.98), epistatic_logic = NULL, prob_of_lpf = list(c(1, 0.97, 0.8,
  0.55), c(1, 0.98, 0.7, 0.51)), barcode_res_pos = list(c(0, 1), c(0,
  2)), prophylactic_pos = list(c(1), c(2)), dur_P = rep(25,
  number_of_drugs), dur_SPC = rep(5, number_of_drugs),
  mft_flag = FALSE, temporal_cycling = -1, sequential_cycling = -1,
  sequential_update = 3, drug_choice = 0,
  partner_drug_ratios = rep(1/number_of_drugs, number_of_drugs))
}
\arguments{
\item{resistance_flag}{Boolean are we simulating resistance}

\item{number_of_drugs}{Numeric for number of drugs used}

\item{number_of_resistance_loci}{Numeric for number of res. loci}

\item{resistance_costs}{Numeric vector for costs of res. Should be 2^res.loci}

\item{epistatic_logic}{Is there compensatory relationships. i.e. what loci 
need to be true for resistance cost to exist. Default of NULL means that 
this becomes seq_len(number_of_resistance_loci), i.e. only dependent on 
their own loci. (TODO: Change this to be a list of length norl)}

\item{prob_of_lpf}{Numeric vector for lpf, Should be 2^res.loci * num_drugs}

\item{mft_flag}{Boolean are we doing mft}

\item{temporal_cycling}{Numeric for when in years a drug switch occurs}

\item{sequential_cycling}{Numeric for what perc. treatment failure before switch}

\item{sequential_update}{How long does it take in years for sequential to be implemented}

\item{drug_choice}{What's the default drug choice to begin. Default = 0}

\item{partner_drug_ratios}{Numeric vector for ratio of first line drugs used}
}
\description{
List for simulating drug usage for resistance/mft variables
}
