% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{drug_list_create}
\alias{drug_list_create}
\title{Create drug list}
\usage{
drug_list_create(resistance_flag = FALSE, number_of_resistance_loci = 3,
  resistance_costs = rep(0.99, 3), prob_of_lpf = rep(0.75, 3),
  mft_flag = FALSE, number_of_drugs = 2, partner_drug_ratios = c(0.5,
  0.5))
}
\arguments{
\item{partner_drug_ratios}{}
}
\description{
List for simulating drug usage for resistance/mft variables
}
