% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_table}
\alias{drug_table}
\title{Drug Efficacy by Genotype Table}
\format{
A data.frame of drug efficacies by genotype
}
\usage{
drug_table
}
\description{
Table of probaility of late parasitological failure (28-day treatment failure)
 per genoytpe. 
 
 Sourced from:
 
 Antimalarial mass drug administration in large populations and the 
 evolution of drug resistance. Tran Dang Nguyen, Thu Nguyen-Anh Tran, 
 Daniel M. Parker, Nicholas J White, Maciej F Boni. 
 bioRxiv 2021.03.08.434496; doi: https://doi.org/10.1101/2021.03.08.434496
}
\keyword{datasets}
