% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_state_create.R
\name{equilibrium_ss_create}
\alias{equilibrium_ss_create}
\title{Equilibrium list creation}
\usage{
equilibrium_ss_create(eqInit)
}
\arguments{
\item{eqInit}{Equilibrium initialisation state as
output from \code{equilibrium_init_create}}
}
\description{
\code{equilibrium_ss_create} creates a starting state from
  the output of \code{equilibrium_init_create}
}
\keyword{internal}
