% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-state-create.R
\name{Equilibrium_SS_Create}
\alias{Equilibrium_SS_Create}
\title{Equilibrium steady state list creation}
\usage{
Equilibrium_SS_Create(eqInit, end.year = 5)
}
\arguments{
\item{eqInit}{Equilibrium initialisation state as output from \code{Equilibrium_Init_Create}}

\item{end.year}{Number of years to run ODE model for. Default = 5.

\code{Model_Param_List_Create}}
}
\description{
\code{Equilibrium_SS_Create} creates a close to steady-state system by 
running an ODE compartment version of the Imperial Malaria model, using
as an initialisation the output of \code{Equilibrium_Init_Create}
}

