% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{housekeeping_list_create}
\alias{housekeeping_list_create}
\title{Create housekeeping parameter list}
\usage{
housekeeping_list_create(quiet = TRUE)
}
\arguments{
\item{quiet}{}
}
\description{
List for simulation housekeeping vars, e.g. quiet prints,
}
