% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_param_lists.R
\name{housekeeping_list_create}
\alias{housekeeping_list_create}
\title{Create housekeeping parameter list}
\usage{
housekeeping_list_create(
  quiet = TRUE,
  quiet_test = TRUE,
  cluster = FALSE,
  clear_up = TRUE
)
}
\arguments{
\item{quiet}{Boolean for quiet simulation. Default = TRUE}

\item{quiet_test}{Boolean for quiet testing statement. Default = TRUE}

\item{cluster}{Boolean for simulation being on cluster. Default = TRUE}

\item{clear_up}{Boolean for whether to clear up the memory used by the 
simulation. Default = TRUE}
}
\description{
List for simulation housekeeping vars, e.g. quiet prints,
}
