% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{intervention_grab}
\alias{intervention_grab}
\title{Intervention grab}
\usage{
intervention_grab(
  country,
  admin,
  year_range,
  final_itn_cov = NULL,
  final_irs_cov = NULL,
  final_ft = NULL
)
}
\arguments{
\item{country}{Country string}

\item{admin}{Admin string}

\item{year_range}{Year range}

\item{final_itn_cov}{Final ITN coverage}

\item{final_irs_cov}{Final IRS coverage}

\item{final_ft}{Final treatment coverage}
}
\description{
Intervention grab
}
\details{
Grabs ITN, IRS  ft from database

N.B. The admin region used for ft can be checked
}
