% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGENTA.R
\docType{package}
\name{MAGENTA}
\alias{MAGENTA}
\alias{MAGENTA-package}
\title{MAGENTA: Individual-based simulation model of malaria epidemiology and genomics.}
\description{
MAGENTA is an individual-based simulation model of malaria epidemiology and genomics. 
MAGENTA extends the imperial malaria model by tracking the infection history of individuals. 
With this additional genetic characteristics of the parasite can be assessed.
}
