% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{mu_fv_create}
\alias{mu_fv_create}
\title{Create mosquito death and feeding vector}
\usage{
mu_fv_create(eqInit, ft = ft, itn_cov = itn_cov, irs_cov = irs_cov,
  int_times = NULL, years = years)
}
\arguments{
\item{years}{}
}
\description{
Grabs spatial incidence and mosquitoFOI vector from database
}
