% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{mu_fv_create}
\alias{mu_fv_create}
\title{Create mosquito death and feeding vector}
\usage{
mu_fv_create(eqInit, ft = ft, itn_cov = itn_cov, irs_cov = irs_cov,
  int_times = NULL, years = years, full = FALSE, complete = TRUE)
}
\arguments{
\item{eqInit}{Equilibrium Inititial Conditions}

\item{ft}{Annual Treatment Coverage vector}

\item{itn_cov}{Annual ITN Coverage vector}

\item{irs_cov}{Annual IRS Coverage vector}

\item{years}{Numeric for total years}

\item{admin}{Admin string}
}
\description{
Grabs spatial incidence and mosquitoFOI vector from database
}
