% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{Param_List_Simulation_Get_Create}
\alias{Param_List_Simulation_Get_Create}
\title{Parameter List creation for MAGENTA simulation getting (saving to disk)}
\usage{
Param_List_Simulation_Get_Create(statePtr)
}
\arguments{
\item{statePtr}{Pointer for current model state as return by \code{Simulation_R}$Ptr}
}
\description{
\code{Param_List_Simulation_Get_Create} creates suitable parameter list for
\code{Simulation_R} for continuing a simulation from memory within the active
session.
}

