% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{Param_List_Simulation_Init_Create}
\alias{Param_List_Simulation_Init_Create}
\title{Parameter List creation for MAGENTA simulation initialisation}
\usage{
Param_List_Simulation_Init_Create(N = 10000, eqSS, barcode_parms,
  spatial_list, housekeeping_list, drug_list, nmf_list,
  vector_adaptation_list)
}
\arguments{
\item{N}{Population size. Default = 1e4}

\item{eqSS}{Output of \code{Equilibrium_Steady_State_Create}}

\item{barcode_parms}{List of barcode/genetic parameters}

\item{spatial_list}{Spatial parmeters to come in}

\item{housekeeping_list}{Housekeeping parameter list 
from \code{housekeeping_list_create}}

\item{drug_list}{Drug parameter list 
from \code{drug_list_create}}

\item{nmf_list}{Non malarial fever parameter list 
from \code{nmf_list_create}}

\item{vector_adaptation_list}{vector adaptation list 
from \code{vector_adaptation_list}}
}
\description{
\code{Param_List_Simulation_Init_Create} creates suitable parameter list for
\code{Simulation_R} for the beginning of a simulation. Also takes an argument
for feeding in spatial parameters/data.
}
