% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{param_list_simulation_init_create}
\alias{param_list_simulation_init_create}
\title{Parameter List creation for magenta simulation initialisation}
\usage{
param_list_simulation_init_create(N = 10000, eqSS, barcode_params,
  spatial_list, housekeeping_list, drug_list, nmf_list,
  vector_adaptation_list)
}
\arguments{
\item{N}{Population size. Default = 1e4}

\item{eqSS}{Output of \code{equilibrium_steady_state_create}}

\item{barcode_params}{List of barcode/genetic parameters}

\item{spatial_list}{Spatial parmeters to come in}

\item{housekeeping_list}{Housekeeping parameter list 
from \code{housekeeping_list_create}}

\item{drug_list}{Drug parameter list 
from \code{drug_list_create}}

\item{nmf_list}{Non malarial fever parameter list 
from \code{nmf_list_create}}

\item{vector_adaptation_list}{vector adaptation list 
from \code{vector_adaptation_list}}
}
\description{
\code{param_list_simulation_init_create} creates suitable parameter list for
\code{simulation_R} for the beginning of a simulation. Also takes an argument
for feeding in spatial parameters/data.
}
