% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{Param_List_Simulation_Saved_Init_Create}
\alias{Param_List_Simulation_Saved_Init_Create}
\title{Parameter List creation for loading saved MAGENTA simulation}
\usage{
Param_List_Simulation_Saved_Init_Create(savedState)
}
\arguments{
\item{savedState}{Saved state generated by \code{Simulation_R} when provided with 
a \code{Param_List_Simulation_Get_Create} parameter list}
}
\description{
\code{Param_List_Simulation_Saved_Init_Create} creates suitable parameter list for
\code{Simulation_R} for continuing a simulation from memory within the active
session.
}
