% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PfPR-to-EIR-heuristic.R
\name{PfPR_to_EIR_heuristic}
\alias{PfPR_to_EIR_heuristic}
\title{Brute force conversion of PfPR and ft to an EIR}
\usage{
PfPR_to_EIR_heuristic(PfPR = NULL, PfPR_micro = NULL, mv = NULL, ft, ...)
}
\arguments{
\item{PfPR}{Parasite prevaence by PCR}

\item{PfPR_micro}{Parasite prevaence by microscopy (Default = NULL)}

\item{mv}{Mosquito size if it's known. Default = NULL}

\item{ft}{Treatment seeking}

\item{...}{Any other params to be fed to the model parameter list}
}
\description{
Brute force conversion of PfPR and ft to an EIR
}
