% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PfPR-to-EIR-heuristic.R
\name{PfPR_to_EIR_heuristic}
\alias{PfPR_to_EIR_heuristic}
\title{Brute force conversion of PfPR and ft to an EIR}
\usage{
PfPR_to_EIR_heuristic(PfPR, ft)
}
\arguments{
\item{PfPR}{Parasite prevaence by PCR}

\item{ft}{Treatment seeking}
}
\description{
Brute force conversion of PfPR and ft to an EIR
}
