% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{Pipeline}
\alias{Pipeline}
\title{Pipeline for cluster submission}
\usage{
Pipeline(EIR = 120, N = 1e+05, years = 20)
}
\arguments{
\item{EIR}{Numeric for desired annual EIR. Default = 120

\code{Pipeline}}

\item{N}{Population Size. Default = 100000}

\item{years}{Lenth of simulation. Default = 10}
}
\description{
\code{Pipeline} steps through creating the parameter list, the equilibrium
initialisation and steady state creation before checking and passing suitable
parameters to the simulation
}

