% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sample_coi}
\alias{sample_coi}
\title{Sample x times from saved simulation and produce age COI data}
\usage{
sample_coi(
  sim_save,
  ID,
  sample_size,
  age_densities,
  age_breaks = seq(0, 90 * 365, 2 * 365),
  sub_patents_included = FALSE,
  reps = 50,
  COI_type = "pcr_imperial"
)
}
\arguments{
\item{sim_save}{Saved output of simulation}

\item{ID}{Vector of detection immunities for the population}

\item{sample_size}{Numeric for sample size}

\item{age_densities}{Densitvy vector for age distribution required}

\item{age_breaks}{Corresponding vector of age breaks for the density}

\item{sub_patents_included}{Boolean as to whether subpatents are included. 
Default = TRUE}

\item{reps}{How many samples are made}

\item{COI_type}{type of COI calculation}
}
\description{
\code{sample_coi} samples from a human save according to some age distirbution
}
