% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seasonal_profile}
\alias{seasonal_profile}
\title{Create seasonal theta return}
\usage{
seasonal_profile(country, admin)
}
\arguments{
\item{country}{Character for country within which admin2 is in. Default = NULL}

\item{admin}{Character for admin region. Some fuzzy logic will be used to match. If 
not provided then no seasonality is introduced. Default = NULL}
}
\description{
Create seasonal theta return
}
