% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-state-create.R
\name{spl_create}
\alias{spl_create}
\title{Create spatial list object to be passed to main functions if required}
\usage{
spl_create(spatial_type = 0, human_importation_rate_vector = NULL,
  mosquito_imporation_rate_vector = NULL,
  cotransmission_freq_vector = rep(1, 10000),
  oocyst_freq_vector = rep(1, 10000), num_human_infs = 0,
  num_mos_infs = 0, plaf = rep(0.5, 24))
}
\description{
\code{spl_create} creates the user required spatial 
list to be passed to the initialisation function if require
}
