% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-state-create.R
\name{spl_create}
\alias{spl_create}
\title{Create spatial list object to be passed to main functions if required}
\usage{
spl_create(island_PLAF = NULL, human_importation_rate_vector = NULL,
  mosquito_imporation_rate_vector = NULL,
  cotransmission_freq_vector = rep(10000, 1),
  oocyst_freq_vector = rep(10000, 1), num_human_infs, num_mos_infs)
}
\description{
\code{spl_create} creates the user required spatial 
list to be passed to the initialisation function if require
}
