% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{summary_data_frames_from_sims}
\alias{summary_data_frames_from_sims}
\title{Create summary prev and genetic dataframes}
\usage{
summary_data_frames_from_sims(res_list, update_length = 30, years = 35,
  age_breaks = list(1, 2, 3), coi_detect = FALSE, ibd = FALSE)
}
\arguments{
\item{res_list}{list of simulation outputs}

\item{update_length}{Days between sim saves in simulation. Default = 30}

\item{years}{Years sims were run for}

\item{ibd}{Were we collecting info on IBD}
}
\description{
Create summary prev and genetic dataframes
}
