% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarySE}
\alias{summarySE}
\title{Create summary statistics df}
\usage{
summarySE(data = NULL, measurevar = "COI", groupvars = c("Age_Bin",
  "State"), conf.interval = 0.95)
}
\arguments{
\item{data}{Dataframe to be summarised}

\item{measurevar}{Character for measure variable}

\item{groupvars}{Characters of grouping variables}

\item{conf.interval}{numeric for CI}
}
\description{
Create summary statistics df
}
