% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_genetics_utils.R
\name{summarySE_mean_only}
\alias{summarySE_mean_only}
\title{Create summary statistics df}
\usage{
summarySE_mean_only(data = NULL, measurevar = "COI",
  groupvars = c("Age_Bin", "State"), mean_only = TRUE)
}
\arguments{
\item{data}{Dataframe to be summarised}

\item{measurevar}{Character for measure variable}

\item{groupvars}{What are we summarise by}

\item{mean_only}{Are we just calculating the mean. Default = TRUE}
}
\description{
Create summary statistics df
}
