% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_genetics_utils.R
\name{summarySE_mean_only_max_mean}
\alias{summarySE_mean_only_max_mean}
\title{Create summary statistics df}
\usage{
summarySE_mean_only_max_mean(
  data = NULL,
  measurevar = "COI",
  groupvars = c("Age_Bin", "State"),
  mean_only = TRUE,
  max = 0
)
}
\arguments{
\item{data}{Dataframe to be summarised}

\item{measurevar}{Character for measure variable}

\item{groupvars}{What are we summarise by}

\item{mean_only}{Are we just calculating the mean. Default = TRUE}

\item{max}{Maximum value that can be obsered in the data beign summarised. 
If greater than zero, all values in data will be set to the max before
summarising. Default = 0.}
}
\description{
Create summary statistics df
}
\keyword{internal}
